+++ src/pages/Nine/components/index.ts
@@ -1,0 +1,2 @@
+export * from './NineGameCard';
+export * from './NineGameDashboard';