+++ src/pages/Nine/index.ts
@@ -1,0 +1,1 @@
+export { Nine as default } from './Nine';